+++ core/src/avm1/globals/xml.rs
@@ -1,573 +1,49 @@
-//! XML/XMLNode global classes
+//! XML class
 
 use crate::avm1::activation::Activation;
 use crate::avm1::error::Error;
 use crate::avm1::object::xml_object::XmlObject;
 use crate::avm1::property_decl::{define_properties_on, Declaration};
-use crate::avm1::{ArrayObject, Object, TObject, Value};
+use crate::avm1::{Object, TObject, Value};
 use crate::avm_warn;
 use crate::backend::navigator::RequestOptions;
 use crate::string::{AvmString, WStr};
-use crate::xml;
 use crate::xml::{XmlDocument, XmlNode};
 use gc_arena::MutationContext;
 use quick_xml::Error as ParseError;
 
-pub const XML_NO_ERROR: f64 = 0.0;
+const XML_NO_ERROR: i32 = 0;
 #[allow(dead_code)]
-pub const XML_CDATA_NOT_TERMINATED: f64 = -2.0;
-pub const XML_DECL_NOT_TERMINATED: f64 = -3.0;
+const XML_CDATA_NOT_TERMINATED: i32 = -2;
+const XML_DECL_NOT_TERMINATED: i32 = -3;
 #[allow(dead_code)]
-pub const XML_DOCTYPE_NOT_TERMINATED: f64 = -4.0;
+const XML_DOCTYPE_NOT_TERMINATED: i32 = -4;
 #[allow(dead_code)]
-pub const XML_COMMENT_NOT_TERMINATED: f64 = -5.0;
-pub const XML_ELEMENT_MALFORMED: f64 = -6.0;
-pub const XML_OUT_OF_MEMORY: f64 = -7.0;
-pub const XML_ATTRIBUTE_NOT_TERMINATED: f64 = -8.0;
+const XML_COMMENT_NOT_TERMINATED: i32 = -5;
+const XML_ELEMENT_MALFORMED: i32 = -6;
+const XML_OUT_OF_MEMORY: i32 = -7;
+const XML_ATTRIBUTE_NOT_TERMINATED: i32 = -8;
 #[allow(dead_code)]
-pub const XML_MISMATCHED_START: f64 = -9.0;
-pub const XML_MISMATCHED_END: f64 = -10.0;
+const XML_MISMATCHED_START: i32 = -9;
+const XML_MISMATCHED_END: i32 = -10;
 
-const XMLNODE_PROTO_DECLS: &[Declaration] = declare_properties! {
-    "localName" => property(xmlnode_local_name; READ_ONLY);
-    "nodeName" => property(xmlnode_node_name; READ_ONLY);
-    "nodeType" => property(xmlnode_node_type; READ_ONLY);
-    "nodeValue" => property(xmlnode_node_value; READ_ONLY);
-    "prefix" => property(xmlnode_prefix; READ_ONLY);
-    "childNodes" => property(xmlnode_child_nodes; READ_ONLY);
-    "firstChild" => property(xmlnode_first_child; READ_ONLY);
-    "lastChild" => property(xmlnode_last_child; READ_ONLY);
-    "parentNode" => property(xmlnode_parent_node; READ_ONLY);
-    "previousSibling" => property(xmlnode_previous_sibling; READ_ONLY);
-    "nextSibling" => property(xmlnode_next_sibling; READ_ONLY);
-    "attributes" => property(xmlnode_attributes; READ_ONLY);
-    "namespaceURI" => property(xmlnode_namespace_uri; READ_ONLY);
-    "appendChild" => method(xmlnode_append_child);
-    "insertBefore" => method(xmlnode_insert_before);
-    "cloneNode" => method(xmlnode_clone_node);
-    "getNamespaceForPrefix" => method(xmlnode_get_namespace_for_prefix);
-    "getPrefixForNamespace" => method(xmlnode_get_prefix_for_namespace);
-    "hasChildNodes" => method(xmlnode_has_child_nodes);
-    "removeNode" => method(xmlnode_remove_node);
-    "toString" => method(xmlnode_to_string);
-};
-
-const XML_PROTO_DECLS: &[Declaration] = declare_properties! {
-    "docTypeDecl" => property(xml_doc_type_decl; READ_ONLY);
+const PROTO_DECLS: &[Declaration] = declare_properties! {
+    "docTypeDecl" => property(doc_type_decl; READ_ONLY);
     "ignoreWhite" => bool(false);
     "contentType" => string("application/x-www-form-urlencoded"; READ_ONLY);
-    "xmlDecl" => property(xml_xml_decl; READ_ONLY);
-    "idMap" => property(xml_id_map; READ_ONLY);
-    "status" => property(xml_status; READ_ONLY);
-    "createElement" => method(xml_create_element);
-    "createTextNode" => method(xml_create_text_node);
-    "parseXML" => method(xml_parse_xml);
-    "load" => method(xml_load);
-    "sendAndLoad" => method(xml_send_and_load);
-    "onData" => method(xml_on_data);
+    "xmlDecl" => property(xml_decl; READ_ONLY);
+    "idMap" => property(id_map; READ_ONLY);
+    "status" => property(status; READ_ONLY);
+    "createElement" => method(create_element);
+    "createTextNode" => method(create_text_node);
+    "parseXML" => method(parse_xml);
+    "load" => method(load);
+    "sendAndLoad" => method(send_and_load);
+    "onData" => method(on_data);
 };
 
-/// Returns true if a particular node can or cannot be exposed to AVM1.
-///
-/// Our internal XML tree representation supports node types that AVM1 XML did
-/// not. Those nodes are filtered from all attributes that return XML nodes to
-/// act as if those nodes did not exist. For example, `prevSibling` skips
-/// past incompatible nodes, etc.
-fn is_as2_compatible(node: &XmlNode<'_>) -> bool {
-    node.is_document_root() || node.is_element() || node.is_text()
-}
-
-/// XMLNode constructor
-pub fn xmlnode_constructor<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    let blank_document = XmlDocument::new(activation.context.gc_context);
-
-    match (
-        args.get(0)
-            .map(|v| v.coerce_to_f64(activation).map(|v| v as u32)),
-        args.get(1).map(|v| v.coerce_to_string(activation)),
-        this.as_xml_node(),
-    ) {
-        (Some(Ok(1)), Some(Ok(ref strval)), Some(ref mut this_node)) => {
-            let mut xmlelement =
-                XmlNode::new_element(activation.context.gc_context, *strval, blank_document);
-            xmlelement.introduce_script_object(activation.context.gc_context, this);
-            this_node.swap(activation.context.gc_context, xmlelement);
-        }
-        (Some(Ok(3)), Some(Ok(ref strval)), Some(ref mut this_node)) => {
-            let mut xmlelement =
-                XmlNode::new_text(activation.context.gc_context, *strval, blank_document);
-            xmlelement.introduce_script_object(activation.context.gc_context, this);
-            this_node.swap(activation.context.gc_context, xmlelement);
-        }
-        //Invalid nodetype ID, string value missing, or not an XMLElement
-        _ => {}
-    };
-
-    Ok(this.into())
-}
-
-pub fn xmlnode_append_child<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let (Some(mut xmlnode), Some(child_xmlnode)) = (
-        this.as_xml_node(),
-        args.get(0)
-            .and_then(|n| n.coerce_to_object(activation).as_xml_node()),
-    ) {
-        if !xmlnode.has_child(child_xmlnode) {
-            let position = xmlnode.children_len();
-            if let Err(e) =
-                xmlnode.insert_child(activation.context.gc_context, position, child_xmlnode)
-            {
-                avm_warn!(
-                    activation,
-                    "Couldn't insert_child inside of XMLNode.appendChild: {}",
-                    e
-                );
-            }
-        }
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn xmlnode_insert_before<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let (Some(mut xmlnode), Some(child_xmlnode), Some(insertpoint_xmlnode)) = (
-        this.as_xml_node(),
-        args.get(0)
-            .and_then(|n| n.coerce_to_object(activation).as_xml_node()),
-        args.get(1)
-            .and_then(|n| n.coerce_to_object(activation).as_xml_node()),
-    ) {
-        if !xmlnode.has_child(child_xmlnode) {
-            if let Some(position) = xmlnode.child_position(insertpoint_xmlnode) {
-                if let Err(e) =
-                    xmlnode.insert_child(activation.context.gc_context, position, child_xmlnode)
-                {
-                    avm_warn!(
-                        activation,
-                        "Couldn't insert_child inside of XMLNode.insertBefore: {}",
-                        e
-                    );
-                }
-            }
-        }
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn xmlnode_clone_node<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let (Some(xmlnode), deep) = (
-        this.as_xml_node(),
-        args.get(0)
-            .map(|v| v.as_bool(activation.swf_version()))
-            .unwrap_or(false),
-    ) {
-        let mut clone_node = xmlnode.duplicate(activation.context.gc_context, deep);
-
-        return Ok(clone_node
-            .script_object(
-                activation.context.gc_context,
-                Some(activation.context.avm1.prototypes.xml_node),
-            )
-            .into());
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn xmlnode_get_namespace_for_prefix<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let (Some(xmlnode), Some(prefix_string)) = (
-        this.as_xml_node(),
-        args.get(0).map(|v| v.coerce_to_string(activation)),
-    ) {
-        if let Some(uri) =
-            xmlnode.lookup_uri_for_namespace(activation.context.gc_context, &prefix_string?)
-        {
-            Ok(uri.into())
-        } else {
-            Ok(Value::Null)
-        }
-    } else {
-        Ok(Value::Undefined)
-    }
-}
-
-pub fn xmlnode_get_prefix_for_namespace<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let (Some(xmlnode), Some(uri_string)) = (
-        this.as_xml_node(),
-        args.get(0).map(|v| v.coerce_to_string(activation)),
-    ) {
-        if let Some(prefix) = xmlnode.lookup_namespace_for_uri(&uri_string?) {
-            Ok(AvmString::new(activation.context.gc_context, prefix).into())
-        } else {
-            Ok(Value::Null)
-        }
-    } else {
-        Ok(Value::Undefined)
-    }
-}
-
-pub fn xmlnode_has_child_nodes<'gc>(
-    _activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let Some(xmlnode) = this.as_xml_node() {
-        Ok((xmlnode.children_len() > 0).into())
-    } else {
-        Ok(Value::Undefined)
-    }
-}
-
-pub fn xmlnode_remove_node<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let Some(node) = this.as_xml_node() {
-        if let Some(mut parent) = node.parent() {
-            if let Err(e) = parent.remove_child(activation.context.gc_context, node) {
-                avm_warn!(activation, "Error in XML.removeNode: {}", e);
-            }
-        }
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn xmlnode_to_string<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let Some(node) = this.as_xml_node() {
-        let result = node.into_string(&mut is_as2_compatible);
-
-        return Ok(AvmString::new_utf8(
-            activation.context.gc_context,
-            result.unwrap_or_else(|e| {
-                avm_warn!(activation, "XMLNode toString failed: {}", e);
-                "".to_string()
-            }),
-        )
-        .into());
-    }
-
-    Ok("".into())
-}
-
-pub fn xmlnode_local_name<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    Ok(this
-        .as_xml_node()
-        .and_then(|n| n.tag_name())
-        .map(|n| AvmString::new(activation.context.gc_context, n.local_name()).into())
-        .unwrap_or_else(|| Value::Null))
-}
-
-pub fn xmlnode_node_name<'gc>(
-    _activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    Ok(this
-        .as_xml_node()
-        .and_then(|n| n.tag_name())
-        .map(|n| Value::from(n.node_name()))
-        .unwrap_or_else(|| Value::Null))
-}
-
-pub fn xmlnode_node_type<'gc>(
-    _activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    Ok(this
-        .as_xml_node()
-        .map(|n| {
-            match n.node_type() {
-                xml::DOCUMENT_NODE => xml::ELEMENT_NODE,
-                xml::DOCUMENT_TYPE_NODE => xml::TEXT_NODE,
-                xml::COMMENT_NODE => xml::TEXT_NODE,
-                n => n,
-            }
-            .into()
-        })
-        .unwrap_or_else(|| Value::Undefined))
-}
-
-pub fn xmlnode_node_value<'gc>(
-    _activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    Ok(this
-        .as_xml_node()
-        .and_then(|n| n.node_value())
-        .map(|v| v.into())
-        .unwrap_or_else(|| Value::Null))
-}
-
-pub fn xmlnode_prefix<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    Ok(this
-        .as_xml_node()
-        .and_then(|n| n.tag_name())
-        .map(|n| {
-            n.prefix()
-                .map(|n| AvmString::new(activation.context.gc_context, n))
-                .unwrap_or_default()
-                .into()
-        })
-        .unwrap_or_else(|| Value::Null))
-}
-
-pub fn xmlnode_child_nodes<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let Some(node) = this.as_xml_node() {
-        return Ok(ArrayObject::new(
-            activation.context.gc_context,
-            activation.context.avm1.prototypes().array,
-            node.children().filter(is_as2_compatible).map(|mut child| {
-                child
-                    .script_object(
-                        activation.context.gc_context,
-                        Some(activation.context.avm1.prototypes.xml_node),
-                    )
-                    .into()
-            }),
-        )
-        .into());
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn xmlnode_first_child<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let Some(node) = this.as_xml_node() {
-        let mut children = node.children();
-        let mut next = children.next();
-        while let Some(my_next) = next {
-            if is_as2_compatible(&my_next) {
-                break;
-            }
-
-            next = my_next.next_sibling();
-        }
-
-        return Ok(next
-            .map(|mut child| {
-                child
-                    .script_object(
-                        activation.context.gc_context,
-                        Some(activation.context.avm1.prototypes.xml_node),
-                    )
-                    .into()
-            })
-            .unwrap_or_else(|| Value::Null));
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn xmlnode_last_child<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let Some(node) = this.as_xml_node() {
-        let mut children = node.children();
-        let mut prev = children.next_back();
-        while let Some(my_prev) = prev {
-            if is_as2_compatible(&my_prev) {
-                break;
-            }
-
-            prev = my_prev.prev_sibling();
-        }
-        return Ok(prev
-            .map(|mut child| {
-                child
-                    .script_object(
-                        activation.context.gc_context,
-                        Some(activation.context.avm1.prototypes.xml_node),
-                    )
-                    .into()
-            })
-            .unwrap_or_else(|| Value::Null));
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn xmlnode_parent_node<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let Some(node) = this.as_xml_node() {
-        return Ok(node
-            .parent()
-            .map(|mut parent| {
-                parent
-                    .script_object(
-                        activation.context.gc_context,
-                        Some(activation.context.avm1.prototypes.xml_node),
-                    )
-                    .into()
-            })
-            .unwrap_or_else(|| Value::Null));
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn xmlnode_previous_sibling<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let Some(node) = this.as_xml_node() {
-        let mut prev = node.prev_sibling();
-        while let Some(my_prev) = prev {
-            if is_as2_compatible(&my_prev) {
-                break;
-            }
-
-            prev = my_prev.prev_sibling();
-        }
-
-        return Ok(prev
-            .map(|mut prev| {
-                prev.script_object(
-                    activation.context.gc_context,
-                    Some(activation.context.avm1.prototypes.xml_node),
-                )
-                .into()
-            })
-            .unwrap_or_else(|| Value::Null));
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn xmlnode_next_sibling<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let Some(node) = this.as_xml_node() {
-        let mut next = node.next_sibling();
-        while let Some(my_next) = next {
-            if is_as2_compatible(&my_next) {
-                break;
-            }
-
-            next = my_next.next_sibling();
-        }
-
-        return Ok(next
-            .map(|mut next| {
-                next.script_object(
-                    activation.context.gc_context,
-                    Some(activation.context.avm1.prototypes.xml_node),
-                )
-                .into()
-            })
-            .unwrap_or_else(|| Value::Null));
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn xmlnode_attributes<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let Some(mut node) = this.as_xml_node() {
-        return Ok(node
-            .attribute_script_object(activation.context.gc_context)
-            .map(|o| o.into())
-            .unwrap_or_else(|| Value::Undefined));
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn xmlnode_namespace_uri<'gc>(
-    activation: &mut Activation<'_, 'gc, '_>,
-    this: Object<'gc>,
-    _args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    if let Some(node) = this.as_xml_node() {
-        if let Some(name) = node.tag_name() {
-            return Ok(node
-                .lookup_uri_for_namespace(
-                    activation.context.gc_context,
-                    name.prefix().unwrap_or_default(),
-                )
-                .unwrap_or_default()
-                .into());
-        }
-
-        return Ok(Value::Null);
-    }
-
-    Ok(Value::Undefined)
-}
-
-/// Construct the prototype for `XMLNode`.
-pub fn create_xmlnode_proto<'gc>(
-    gc_context: MutationContext<'gc, '_>,
-    proto: Object<'gc>,
-    fn_proto: Object<'gc>,
-) -> Object<'gc> {
-    let xmlnode_proto = XmlObject::empty_node(gc_context, Some(proto));
-    let object = xmlnode_proto.as_script_object().unwrap();
-    define_properties_on(XMLNODE_PROTO_DECLS, gc_context, object, fn_proto);
-    xmlnode_proto
-}
-
 /// XML (document) constructor
-pub fn xml_constructor<'gc>(
+pub fn constructor<'gc>(
     activation: &mut Activation<'_, 'gc, '_>,
     this: Object<'gc>,
     args: &[Value<'gc>],
@@ -611,7 +87,7 @@ pub fn xml_constructor<'gc>(
     Ok(this.into())
 }
 
-pub fn xml_create_element<'gc>(
+fn create_element<'gc>(
     activation: &mut Activation<'_, 'gc, '_>,
     this: Object<'gc>,
     args: &[Value<'gc>],
@@ -638,7 +114,7 @@ pub fn xml_create_element<'gc>(
     Ok(object.into())
 }
 
-pub fn xml_create_text_node<'gc>(
+fn create_text_node<'gc>(
     activation: &mut Activation<'_, 'gc, '_>,
     this: Object<'gc>,
     args: &[Value<'gc>],
@@ -665,7 +141,7 @@ pub fn xml_create_text_node<'gc>(
     Ok(object.into())
 }
 
-pub fn xml_parse_xml<'gc>(
+fn parse_xml<'gc>(
     activation: &mut Activation<'_, 'gc, '_>,
     this: Object<'gc>,
     args: &[Value<'gc>],
@@ -708,7 +184,7 @@ pub fn xml_parse_xml<'gc>(
     Ok(Value::Undefined)
 }
 
-pub fn xml_send_and_load<'gc>(
+fn send_and_load<'gc>(
     activation: &mut Activation<'_, 'gc, '_>,
     this: Object<'gc>,
     args: &[Value<'gc>],
@@ -731,7 +207,7 @@ pub fn xml_send_and_load<'gc>(
     Ok(Value::Undefined)
 }
 
-pub fn xml_load<'gc>(
+fn load<'gc>(
     activation: &mut Activation<'_, 'gc, '_>,
     this: Object<'gc>,
     args: &[Value<'gc>],
@@ -752,7 +228,7 @@ pub fn xml_load<'gc>(
     }
 }
 
-pub fn xml_on_data<'gc>(
+fn on_data<'gc>(
     activation: &mut Activation<'_, 'gc, '_>,
     this: Object<'gc>,
     args: &[Value<'gc>],
@@ -773,14 +249,14 @@ pub fn xml_on_data<'gc>(
     Ok(Value::Undefined)
 }
 
-pub fn xml_doc_type_decl<'gc>(
+fn doc_type_decl<'gc>(
     activation: &mut Activation<'_, 'gc, '_>,
     this: Object<'gc>,
     _args: &[Value<'gc>],
 ) -> Result<Value<'gc>, Error<'gc>> {
     if let Some(node) = this.as_xml_node() {
         if let Some(doctype) = node.document().doctype() {
-            let result = doctype.into_string(&mut |_| true);
+            let result = doctype.into_string(&|_| true);
 
             return Ok(AvmString::new_utf8(
                 activation.context.gc_context,
@@ -796,7 +272,7 @@ pub fn xml_doc_type_decl<'gc>(
     Ok(Value::Undefined)
 }
 
-pub fn xml_xml_decl<'gc>(
+fn xml_decl<'gc>(
     activation: &mut Activation<'_, 'gc, '_>,
     this: Object<'gc>,
     _args: &[Value<'gc>],
@@ -818,7 +294,7 @@ pub fn xml_xml_decl<'gc>(
     Ok(Value::Undefined)
 }
 
-pub fn xml_id_map<'gc>(
+fn id_map<'gc>(
     activation: &mut Activation<'_, 'gc, '_>,
     this: Object<'gc>,
     _args: &[Value<'gc>],
@@ -833,7 +309,7 @@ pub fn xml_id_map<'gc>(
     Ok(Value::Undefined)
 }
 
-pub fn xml_status<'gc>(
+fn status<'gc>(
     _activation: &mut Activation<'_, 'gc, '_>,
     this: Object<'gc>,
     _args: &[Value<'gc>],
@@ -873,7 +349,7 @@ fn spawn_xml_fetch<'gc>(
     let request_options = if let Some(node) = send_object {
         // Send `node` as string
         RequestOptions::post(Some((
-            node.into_string(&mut is_as2_compatible)
+            node.into_string(&XmlNode::is_as2_compatible)
                 .unwrap_or_default()
                 .into_bytes(),
             "application/x-www-form-urlencoded".to_string(),
@@ -912,13 +388,13 @@ fn spawn_xml_fetch<'gc>(
 }
 
 /// Construct the prototype for `XML`.
-pub fn create_xml_proto<'gc>(
+pub fn create_proto<'gc>(
     gc_context: MutationContext<'gc, '_>,
     proto: Object<'gc>,
     fn_proto: Object<'gc>,
 ) -> Object<'gc> {
     let xml_proto = XmlObject::empty_node(gc_context, Some(proto));
     let object = xml_proto.as_script_object().unwrap();
-    define_properties_on(XML_PROTO_DECLS, gc_context, object, fn_proto);
+    define_properties_on(PROTO_DECLS, gc_context, object, fn_proto);
     xml_proto
 }
